METRON-309 Altered based on Stellar Unification changes

+++ metron-analytics/metron-profiler/src/main/java/org/apache/metron/profiler/bolt/ProfileBuilderBolt.java
@@ -199,7 +199,7 @@ public class ProfileBuilderBolt extends ConfiguredProfilerBolt {
 
     // execute the 'result' expression
     String resultExpr = profileConfig.getResult();
-    Long result = executor.execute(resultExpr, new JSONObject(), Long.class);
+    Double result = executor.execute(resultExpr, new JSONObject(), Double.class);
 
     // emit the completed profile measurement
     measurement.setEnd(getTimestamp());
